pyproject.toml
[tool.poetry]
name = "auth-service"
version = "1.0.0"
description = "Authenticaion service"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
dependency-injector = "^4.41.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.19"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
loguru = "^0.7.2"
alembic = "^1.13.2"
pytz = "^2024.1"
bcrypt = "^4.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
